Remove redux from Login Options screen
